requirements.txt
matplotlib==3.8.2
numpy==1.26.2
pandas==2.1.3
Pillow==10.1.0
Pillow==10.1.0
scikit_learn==1.3.2
torch==2.1.1+cu118
torchvision==0.16.1
tqdm==4.66.1
umap==0.1.1